+++ util/shared/modules.mjs
@@ -54,8 +54,11 @@ const getMethods = compose(
     getPath(['methods'])
 )
 
-const isProviderMethod = compose(
+const isProviderInterfaceMethod = compose(
     and(
+        compose(
+            propSatisfies('name', name => name.startsWith('onRequest'))
+        ),
         compose(
             option(false),
             map(_ => true),
@@ -68,8 +71,7 @@ const isProviderMethod = compose(
               )
             ),
             getPath(['tags']),        
-        ),
-        propSatisfies('name', (prop) => prop.startsWith('onRequest'))
+        )
     )
   )
 
@@ -87,11 +89,11 @@ const getProvidedCapabilitiesFromMethod = (json) => {
 }
 
 const getProvidedCapabilities = (json) => {
-    return Array.from(new Set([...getMethods(json).filter(isProviderMethod).map(method => method.tags.find(tag => tag['x-provides'])['x-provides'])]))
+    return Array.from(new Set([...getMethods(json).filter(isProviderInterfaceMethod).map(method => method.tags.find(tag => tag['x-provides'])['x-provides'])]))
 }
 
-const getMethodsThatProvide = (capability, json) => {
-    return getMethods(json).filter(method => method.tags && method.tags.find(tag => tag['x-provides'] === capability))
+const getProviderInterfaceMethods = (capability, json) => {
+    return getMethods(json).filter(method => method.name.startsWith("onRequest") && method.tags && method.tags.find(tag => tag['x-provides'] === capability))
 }
   
 
@@ -490,13 +492,8 @@ const createFocusFromProvider = provider => {
     const ready = JSON.parse(JSON.stringify(provider))
     ready.name = ready.name.charAt(9).toLowerCase() + ready.name.substr(10) + 'Focus'
     ready.summary = `Internal API for ${provider.name.substr(9)} Provider to request focus for UX purposes.`
-    const old_tags = ready.tags
-    ready.tags = [
-        {
-            'name': 'rpc-only',
-            'x-allow-focus-for': provider.name
-        }
-    ]
+    ready.tags = ready.tags.filter(t => t.name !== 'event')
+    ready.tags.find(t => t.name === 'capabilities')['x-allow-focus-for'] = provider.name
 
     ready.params = []
     ready.result = {
@@ -530,13 +527,9 @@ const createResponseFromProvider = (provider, type, json) => {
     const response = JSON.parse(JSON.stringify(provider))
     response.name = response.name.charAt(9).toLowerCase() + response.name.substr(10) + type
     response.summary = `Internal API for ${provider.name.substr(9)} Provider to send back ${type.toLowerCase()}.`
-    const old_tags = response.tags
-    response.tags = [
-        {
-            'name': 'rpc-only'
-        }
-    ]
-    response.tags[`x-${type.toLowerCase()}-for`] = provider.name
+
+    response.tags = response.tags.filter(t => t.name !== 'event')
+    response.tags.find(t => t.name === 'capabilities')[`x-${type.toLowerCase()}-for`] = provider.name
 
     const paramExamples = []
 
@@ -718,7 +711,7 @@ const generateTemporalSetMethods = json => {
 
 
 const generateProviderMethods = json => {
-    const providers = json.methods.filter(isProviderMethod) || []// m => m.tags && m.tags.find( t => t.name == 'capabilities' && t['x-provides'])) || []
+    const providers = json.methods.filter(isProviderMethod) || []
 
     providers.forEach(provider => {
         if (! isRPCOnlyMethod(provider)) {
@@ -837,13 +830,13 @@ export {
     isTemporalSetMethod,
     isExcludedMethod,
     isRPCOnlyMethod,
-    isProviderMethod,
+    isProviderInterfaceMethod,
     hasExamples,
     hasTitle,
     hasMethodAttributes,
     getMethodAttributes,
     getMethods,
-    getMethodsThatProvide,
+    getProviderInterfaceMethods,
     getProvidedCapabilities,
     getEnums,
     getTypes,